Version 1.3.1
git-svn-id: https://plugins.svn.wordpress.org/akeebabackupcore/trunk@1174518 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ app/Solo/Model/Manage.php
@@ -205,7 +205,10 @@ class Manage extends Model
 		$allFiles = Factory::getStatistics()->get_all_filenames($stat, false);
 
 		// Remove the custom log file if necessary
-		$this->_deleteLogs($stat);
+		if (!is_null($stat))
+		{
+			$this->_deleteLogs($stat);
+		}
 
 		// Make sure we have some files
 		if (empty($allFiles))
